requirements.txt
llama-cpp-python
fastapi
uvicorn
python-docx
PyPDF2
beautifulsoup4
markdown
pyyaml
striprtf
numpy
langchain
langchain-community
llama-index
llama-index-llms-langchain
llama-index-embeddings-langchain